fix: compressed option should be false by default (#567)
The `compressed` option is not enabled by default. There is no code
setting it to true, except where the old `gzip` property is true. I
confirmed this behavior in testing. The documentation for it appears to
have been inadvertently updated when the `timing` option was added in
commit 1d26bb8.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Documentation**
- Updated `README.md` to reflect the change in default value for the
`compressed` option in the `async request(url[, options])` method from
`true` to `false`.
  
- **Bug Fixes**
- Changed the default behavior of the `compressed` property in request
handling to prevent automatic decoding of `gzip` and `br` responses
unless specified by the user.

- **Chores**
- Deprecated the `ctx` property in `RequestOptions`, recommending the
use of the `opaque` property instead.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->